If Else Logics Test
If I go to Imtiaz by bike 🏍, I will buy 10 products. If the bill amounts to Rs. 5000, I will get a 25% discount. If it amounts to Rs. 10,000, I will get a 50% discount. If it amounts to Rs. 25,000, I will get a 75% discount.

If I do not go to Imtiaz by bike 🏍, I will buy 20 products. If the bill amounts to Rs. 20,000, I will get a 10% discount. If it amounts to Rs. 30,000, I will get a 50% discount. If it amounts to Rs. 75,000, I will get a 75% discount.